Fixed a bug with projectiles being stuck in overlapping time walls
So the problem was that if there was a projectile in a time wall, that then was put into another time wall, the second time wall would percieve the projectile's velocity to be 0, so this change just creates a variable for projectiles called current velocity, whose implementation fixes the aforementioned problem.  Also, I have tested this with multiple different time players by myself, and there seemed to be no issues with the changes.  This should be the last bug that needed fixing for the time wall ability.