Finished PoisonArrow's direcionality
It now points in the direction it is going and works as intended.  However there was a bug regarding it's speed when released from the time-bubble.  That problem is currently believed to be on the side of the time wizard as it seems to appear for fireballs as well as poison arrows.