Update to bambooSpawner
All abilities for Druid now work, this update only makes it so that bamboo can only be spawned while the player is on the ground.
It also decreases the cooldown time of the poison arrow from 1s to .5s, as formerly it was nearly impossible to actually use the druid's spells fast enough to run out of magic.