Implemented fire mine's explosion
With this, the fire player's new abilities are ready for play-testing!!!

Fire mine now works as intended, and causes an explosion when stepped on by an opponent.

Note -possible error between server & client sides, where sometimes the mine will explode for the client, and not damage them, but not explode for the server, then can be set off once again, this time for both sides, however only damaging the clientside player...  I think.  If so, I've no idea what to do about this.