Fixed fire mine bug
Fixed the problem where the fire mine would at first only paritally detonate for the client, and not do anything for the server/other clients, then if the client went to the space where the mine was, it would damage them and it would explode for the others.  Fixed by changing the mine's collider from a plane to a sphere...  Idk why that fixed it, but it does.